requirements.txt
# --- 核心科学计算库 (版本锁死，严禁升级) ---
numpy==1.22.4
scipy==1.9.1
pandas
matplotlib

# --- 深度学习辅助 ---
tensorflow==2.11.0
tensorboard
transformers==4.26.1
albumentations==1.3.0
einops==0.6.0
omegaconf==2.3.0
tqdm

# --- 贝叶斯优化与高斯过程 (ROOT 依赖) ---
botorch==0.6.4
gpytorch==1.6.0

# --- 强化学习与机器人环境 (最难装的部分) ---
# 注意：gym 必须是 0.12.5，否则 design-bench 报错
gym==0.12.5
transforms3d
# robel 和 morphing_agents 放在这里，稍后用特殊命令安装

# --- 离线优化基准 ---
design-bench==2.0.12

# --- 你的 SP-RFM 算法需要 ---
pot